feat: generics and static reflection

+++ SharedPowerUps/SharedPowerUp.cs
@@ -6,6 +6,6 @@ namespace SharedPowerUps;
 public partial class SharedPowerUp : Node
 {
 #pragma warning disable CA1822 // annoying lint about making this static :P
-  public void OnSharedPowerUp(long _) { }
+  public void OnSharedPowerUp(int _) { }
 #pragma warning restore CA1822
 }
